Use overflowing add, sub, and shifts in ALU simulator.
Rust in debug mode panics at overflows, which is very awesome. But we're
emulating an ALU, so I need to choose a hardware behavior and emulate that. I
used integer's `overflowing_{add,bug,shl,shr}` methods, which does the standard
2's complement overflow behavior and also returns a boolean indicating if an
overlow happened or not.

I found the following article very helpful: http://huonw.github.io/blog/2016/04/myths-and-legends-about-integer-overflow-in-rust/